Refactor logo related components into one folder
