+++ src/ui/logo/Logo.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { create } from 'react-test-renderer';
 
-import Logo from 'src/ui/Logo';
+import Logo from 'src/ui/logo/Logo';
 
 it('renders without crashing', () => {
   const renderer = create(<Logo />);
