+++ src/ui/logo/Logo.tsx
@@ -1,28 +1,28 @@
-import NysseLogo from "src/ui/NysseLogo";
-import HslLogo from "src/ui/HslLogo";
 import * as React from 'react';
-import { render } from 'react-dom';
-import MatkaLogo from 'src/ui/MatkaLogo';
-import LinkkiLogo from 'src/ui/LinkkiLogo';
+
+import NysseLogo from 'src/ui/logo/NysseLogo';
+import HslLogo from 'src/ui/logo/HslLogo';
+import MatkaLogo from 'src/ui/logo/MatkaLogo';
+import LinkkiLogo from 'src/ui/logo/LinkkiLogo';
 
 interface ICommonProps {
   readonly monitorConfig?: any,
 }
-class Logo extends React.Component<ICommonProps> {
 
+class Logo extends React.Component<ICommonProps> {
   public render() {
    const monitorConfig = (this.props as ICommonProps).monitorConfig
 
     if(monitorConfig) {
       const feedId = monitorConfig.feedId;
       switch (feedId) {
-        case "tampere":
+        case 'tampere':
           return <NysseLogo />;
-        case "hsl":
+        case 'hsl':
           return <HslLogo />;
-        case "matka":
+        case 'matka':
           return <MatkaLogo />;
-        case "linkki":
+        case 'linkki':
           return <LinkkiLogo />;
         default:
           return null;
@@ -31,4 +31,5 @@ class Logo extends React.Component<ICommonProps> {
     return null;
   }
 }
+
 export default Logo;
